This is checking in what will be 3.29.4beta.
Primary fixes so far:

  Several changes to configure.ac to eliminate all reference to libaes
and to set macros ENDIAN_BIG or ENDIAN_LITTLE to 1 in the configure
stage of the build.  I plan to insert a very simple prequel in Brian
Goodman's brg_endian.h header file that handles endian issues cleanly
and skips most of the stuff below for little endian.  I do need to
ensure that it builds on i386 as well (when I'm done) as I have a report
that 3beta doesn't build clean on that architecture due to problems in
this header file.

  A fix due to Glenn Emelko, GEmelko@aclara.com, where I correctly
bumped filecount to type off_t in libdieharder.h but failed to redefine
rtot and rptr accordingly in the rng_file_input.c struct and code.  He
was running 18 GB raw files and obviously this overflowed uint variables
with bad results.  Oops, and thanks Glenn.

  I am trying to get sts_serial.h to run at 24 bits by default, not 16
(I think that this will still take a not unreasonable amount of time).
The problem is that sts_serial doesn't use bits.c calls to parse out the
next 24 bits, it just grabs 16 bit, then the next 16 bits, out of 32 bit
uints.  This is fast but not scalable.  I have to go in and edit the
code to use bits calls to get the next 24 bits, no foolin, or better yet
use -n ntuple to set the maximum number of bits teste (that's my real
goal, with 24 being the default).

Plans:  David Bauer sent me a fairly extensive patch against 3.29.3beta
that fixes some memory leaks and/or speed issues in bits (?) as well as
fixing some parts of the diehard OP code -- probably fixing Marsaglia's
bugs and not mine, but hard to say.  There are bugs in there and I've
already squashed several so it wouldn't surprise me if there are more
(even more of my own:-).  I'm going to TRY to implement most of his
fixes if they work well and seem to fix something that makes sense to
me, although I'd feel better (per fix) if I could find a test case that
illustrates the failure.  I may have to ask him how he found the bugs so
I can document them in svn somewhere, later.  Memory leaks of course are
relatively easy and again, I could easily have created some -- getting
rid of them is definitely called for.  David is also looking at the
rgb_bitdist tests (which SHOULD be as sensitive as the OP tests if
cranked up to the correct degree) -- there may be some fixes there
coming.  

Finally, I have a few operational changes in mind -- primarily adding or
altering the new interface in a couple of small ways.  Kuiper will go
away as an option (but not the code -- I'll leave a macro in place that
can switch it back on in case there is ever any point in reconsidering
the test, if for example I or David or somebody else figures out an
exact CDF for it so it becomes as accurate and perhaps faster than KS,
or it is needed for a specific rng test in the future.  -k flags will be
used to control how hard ks works (and hence how fast vs accurate
dieharder is) with a default of pretty fast, pretty accurate and
alternatives of slow but to-convergence-exact and really fast but only
accurate enough for the short version of the -a(ll) run.

I'd also like to introduce two new run modes controlled by flags.  One
of them, -m(ultiplier), will allow the user to enter a scale factor to
be applied to the default 100 -psamples used in -a(ll) runs (otherwise
ignored).  So if one want to run all the tests but with 1000 psamples
per test (or 10x whatever the per-test default is) one runs -a -m 10.
This should make it MUCH easier to test to destruction, increase test
resolution, etc.  

Second, I want to introduce a flag that runs a test "to failure" --
something I've planned to do for a long time.  David has already hacked
in his own version of doing this, and I used to do something very
similar in my numerical simulations.  The idea of running in -x(treme)
mode (or whatever I name the flag) would be to start with e.g. 100
pvalues and then add 100 pvalues at a time to the test run until the
final pvalue fails a fairly stringent (user selectable) cutoff.

  -d 1 -x 0.000001

would add psamples to the birthdays test until the final pvalue is under
0.000001.

  -d 1 -X 0.000001

would do the same thing, but it would run the test to this degree of
failure psamples = 100 times with different rng seeds (if appropriate)
and return something like max, min, mean number of psamples required to
cause a test failure.  SOMETHING like this is going to be needed,
because I think it is entirely plausible that some tests have
"poisonous" seeds that have just the right prime modulus to introduce
correlations in their stream, but that NEARLY ALWAYS are started with
seeds that yield good streams.

I'd like to have these last two options work for -a -m runs as well, so

  -a -m 10 -X 0.000001

runs all tests until they fail low or high at one part in a million,
1000 times for different seeds per run, returning the average number of
psamples required to reach failure.  I'd even like to be able to plot
the distribution of this number so one can pick out e.g. bimodal
distributions (bad seeds!) etc.

At some point being able to do everything that dieharder will want/need
to do is going to require a GUI -- something that can generate scatter
plots, candlesticks, real non-ascii histograms, line graphs, 2d/3d
surfaces.  But that's still a ways in the future.  -X is going to be
pretty tricky as well, as dieharder isn't equipped to return anything
but a final cumulative "pvalue" in [0,1] for a test.  But it is probably
better to do it now in the beta phase where this doesn't really damage
any other future dependent interfaces (e.g R).

+++ include/dieharder/libdieharder.h
@@ -199,7 +199,7 @@
   * file_input_raw.
   */
  uint file_input_get_rewind_cnt(gsl_rng *rng);
- uint file_input_get_rtot(gsl_rng *rng);
+ off_t file_input_get_rtot(gsl_rng *rng);
  void file_input_set_rtot(gsl_rng *rng,uint value);
 
  char filename[K];      /* Input file name */
@@ -226,8 +226,8 @@
  typedef struct {
     FILE *fp;
     off_t flen;
-    uint rptr;
-    uint rtot;
+    off_t rptr;
+    off_t rtot;
     uint rewind_cnt;
   } file_input_state_t;
