+++ libdieharder/rng_file_input_raw.c
@@ -28,7 +28,7 @@ static void file_input_raw_set (void *vstate, unsigned long int s);
  *   fp is the file pointer
  *   flen is the number of rands in the file (filecount)
  *   rptr is a count of rands returned since last rewind
- *   rtot is  * a count of rands returned since the file was opened or it
+ *   rtot is a count of rands returned since the file was opened or it
  *      was deliberately reset.
  *   rewind_cnt is a count of how many times the file was rewound since
  *      its last open.
@@ -219,8 +219,8 @@ static void file_input_raw_set (void *vstate, unsigned long int s)
      state->rptr = 0;
      state->rewind_cnt++;
      if(verbose == D_FILE_INPUT_RAW || verbose == D_ALL){
-       fprintf(stderr,"# file_input_raw(): Rewinding %s at rtot = %u\n", filename,state->rtot);
-       fprintf(stderr,"# file_input_raw(): Rewind count = %u, resetting rptr = %u\n",state->rewind_cnt,state->rptr);
+       fprintf(stderr,"# file_input_raw(): Rewinding %s at rtot = %u\n", filename,(uint) state->rtot);
+       fprintf(stderr,"# file_input_raw(): Rewind count = %u, resetting rptr = %u\n",state->rewind_cnt,(uint) state->rptr);
      }
    } else {
      return;
